pyproject.toml
[tool.poetry]
name = "guava_preconditions"
version = "0.1.1"
description = "A python version of guava's preconditions"
authors = ["Elliana <me@mause.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
pytest = "^6.2.5"
pre-commit = "^2.15.0"
pdoc3 = "^0.10.0"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
